+++ wgpu/src/util/mod.rs
@@ -7,6 +7,7 @@ mod belt;
 mod device;
 mod encoder;
 mod init;
+mod texture_blitter;
 
 use std::sync::Arc;
 use std::{
@@ -19,19 +20,12 @@ pub use belt::StagingBelt;
 pub use device::{BufferInitDescriptor, DeviceExt};
 pub use encoder::RenderEncoder;
 pub use init::*;
+pub use texture_blitter::TextureBlitter;
 pub use wgt::{
     math::*, DispatchIndirectArgs, DrawIndexedIndirectArgs, DrawIndirectArgs, TextureDataOrder,
 };
 
-use crate::{
-    dispatch, AddressMode, BindGroupDescriptor, BindGroupEntry, BindGroupLayout,
-    BindGroupLayoutDescriptor, BindGroupLayoutEntry, BindingType, ColorTargetState, ColorWrites,
-    CommandEncoder, Device, Face, FilterMode, FragmentState, FrontFace, LoadOp, MultisampleState,
-    PipelineCompilationOptions, PipelineLayoutDescriptor, PrimitiveState, PrimitiveTopology,
-    RenderPassDescriptor, RenderPipeline, RenderPipelineDescriptor, Sampler, SamplerBindingType,
-    SamplerDescriptor, ShaderModuleDescriptor, ShaderSource, ShaderStages, StoreOp, TextureFormat,
-    TextureSampleType, TextureView, TextureViewDimension, VertexState,
-};
+use crate::dispatch;
 
 /// Treat the given byte slice as a SPIR-V module.
 ///
@@ -196,184 +190,5 @@ pub fn pipeline_cache_key(adapter_info: &wgt::AdapterInfo) -> Option<String> {
             adapter_info.vendor, adapter_info.device
         )),
         _ => None,
-    }
-}
-
-/// A texture blitting utility
-pub struct TextureBlitter {
-    pipeline: RenderPipeline,
-    bind_group_layout: BindGroupLayout,
-    sampler: Sampler,
-}
-
-impl TextureBlitter {
-    /// Returns a new [`TextureBlitter`]
-    /// # Arguments
-    /// - `device` - A [`Device`]
-    /// - `format` - The [`TextureFormat`] of the texture that will be copied to
-    /// - `sample_type` - The [`Sampler`] Filtering Mode
-    pub fn new(device: Device, format: TextureFormat, sample_type: FilterMode) -> Self {
-        let sampler = device.create_sampler(&SamplerDescriptor {
-            label: Some("wgpu::util::TextureBlitter::sampler"),
-            address_mode_u: AddressMode::ClampToEdge,
-            address_mode_v: AddressMode::ClampToEdge,
-            address_mode_w: AddressMode::ClampToEdge,
-            mag_filter: sample_type,
-            ..Default::default()
-        });
-
-        let bind_group_layout = device.create_bind_group_layout(&BindGroupLayoutDescriptor {
-            label: Some("wgpu::util::TextureBlitter::bind_group_layout"),
-            entries: &[
-                BindGroupLayoutEntry {
-                    binding: 0,
-                    visibility: ShaderStages::FRAGMENT,
-                    ty: BindingType::Texture {
-                        sample_type: TextureSampleType::Float { filterable: false },
-                        view_dimension: TextureViewDimension::D2,
-                        multisampled: false,
-                    },
-                    count: None,
-                },
-                BindGroupLayoutEntry {
-                    binding: 1,
-                    visibility: ShaderStages::FRAGMENT,
-                    ty: BindingType::Sampler(SamplerBindingType::Filtering),
-                    count: None,
-                },
-            ],
-        });
-
-        let pipeline_layout = device.create_pipeline_layout(&PipelineLayoutDescriptor {
-            label: Some("wgpu::util::TextureBlitter::pipeline_layout"),
-            bind_group_layouts: &[&bind_group_layout],
-            push_constant_ranges: &[],
-        });
-
-        let shader = device.create_shader_module(ShaderModuleDescriptor {
-            label: Some("wgpu::util::TextureBlitter::shader"),
-            source: ShaderSource::Wgsl(Cow::Borrowed(
-                r#"
-                struct VertexOutput {
-                    @builtin(position) position: vec4<f32>,
-                    @location(0) tex_coords: vec2<f32>,
-                }
-
-                @vertex
-                fn vs_main(@builtin(vertex_index) vi: u32) -> VertexOutput {
-                    var out: VertexOutput;
-
-                    out.tex_coords = vec2<f32>(
-                        f32(vi  << 1u),
-                        f32(vi & u2),
-                    );
-
-                    out.position = vec4<f32>(out.uv * 2.0 - 1.0, 0.0, 1.0);
-
-                   // Invert y so the texture is not upside down
-                   out.tex_coords.y = 1.0 - out.tex_coords.y;
-                   return out;
-                  }
-
-                 @group(0) @binding(0)
-                 var texture: texture_2d<f32>;
-                 @group(0) @binding(0)
-                 var texture_sampler: Sampler;
-
-                 @fragment
-                 fn fs_main(vs: VertexOutput) -> @location(0) vec4<f32> {
-                    return textureSample(texture, texture_sampler, vs.uv);
-                 }
-                "#,
-            )),
-        });
-        let pipeline = device.create_render_pipeline(&RenderPipelineDescriptor {
-            label: Some("wgpu::uti::TextureBlitter::pipeline"),
-            layout: Some(&pipeline_layout),
-            vertex: VertexState {
-                module: &shader,
-                entry_point: Some("vs_main"),
-                compilation_options: PipelineCompilationOptions::default(),
-                buffers: &[],
-            },
-            primitive: PrimitiveState {
-                topology: PrimitiveTopology::TriangleList,
-                strip_index_format: None,
-                front_face: FrontFace::Ccw,
-                cull_mode: Some(Face::Back),
-                unclipped_depth: false,
-                polygon_mode: wgt::PolygonMode::Fill,
-                conservative: false,
-            },
-            depth_stencil: None,
-            multisample: MultisampleState::default(),
-            fragment: Some(FragmentState {
-                module: &shader,
-                entry_point: Some("fs_main"),
-                compilation_options: PipelineCompilationOptions::default(),
-                targets: &[Some(ColorTargetState {
-                    format,
-                    blend: None,
-                    write_mask: ColorWrites::ALL,
-                })],
-            }),
-            multiview: None,
-            cache: None,
-        });
-
-        Self {
-            pipeline,
-            bind_group_layout,
-            sampler,
-        }
-    }
-
-    /// Copies the data from the source [`TextureView`] to the target [`TextureView`]
-    ///
-    /// # Arguments
-    /// - `device` - A [`Device`]
-    /// - `encoder` - A [`CommandEncoder`]
-    /// - `source` - A [`TextureView`] that gets copied from
-    /// - `target` - A [`TextureView`] that gets the data copied from the `source`
-    pub fn copy(
-        &self,
-        device: &Device,
-        encoder: &mut CommandEncoder,
-        source: &TextureView,
-        target: &TextureView,
-    ) {
-        let bind_group = device.create_bind_group(&BindGroupDescriptor {
-            label: Some("wgpu::util::TextureBlitter::bind_group"),
-            layout: &self.bind_group_layout,
-            entries: &[
-                BindGroupEntry {
-                    binding: 0,
-                    resource: crate::BindingResource::TextureView(source),
-                },
-                BindGroupEntry {
-                    binding: 1,
-                    resource: crate::BindingResource::Sampler(&self.sampler),
-                },
-            ],
-        });
-
-        let mut pass = encoder.begin_render_pass(&RenderPassDescriptor {
-            label: Some("wgpu::util::TextureBlitter::pass"),
-            color_attachments: &[Some(crate::RenderPassColorAttachment {
-                view: target,
-                resolve_target: None,
-                ops: wgt::Operations {
-                    load: LoadOp::Load,
-                    store: StoreOp::Store,
-                },
-            })],
-            depth_stencil_attachment: None,
-            timestamp_writes: None,
-            occlusion_query_set: None,
-        });
-
-        pass.set_pipeline(&self.pipeline);
-        pass.set_bind_group(0, &bind_group, &[]);
-        pass.draw(0..3, 0..1);
     }
 }
